new: notifications repo

+++ domain/model/notification/notification.go
@@ -8,12 +8,12 @@ import (
 const DefaultScheduleTime = "10:00:00"
 
 type Notification struct {
-	Id           int
+	Id           int64
 	ScheduledFor time.Time
 	Enable       bool
 	User         *user.User
 }
 
-func New(id int, schedule time.Time, enable bool, user *user.User) *Notification {
+func New(id int64, schedule time.Time, enable bool, user *user.User) *Notification {
 	return &Notification{id, schedule, enable, user}
 }
